refactor post compose page
