+++ backend/src/resource/asset/index.html
@@ -2,6 +2,7 @@
         <meta charset="utf-8">
         <title>70年代、80年代、90年代，怀旧、复古、回忆</title>
         <link rel="stylesheet" href="/asset/purecss.min-b565cdb90b7aac9f.css">
+        <link rel="stylesheet" href="/asset/tag-4d2c4d8d06ebde6.css">
         <style type="text/css">
             body{padding:20px}
             .tar{text-align:right}
@@ -9,6 +10,6 @@
             .tag-btn{color:white;border-radius:4px;text-shadow:0 1px 1px rgba(0, 0, 0, 0.2);background:rgb(28, 184, 65);margin-right:6px}
             .tag-link{margin-right:6px}
         </style>
-    <script type="module">import init from '/asset/index-5114c039e9a0540e.js';init('/asset/index-5114c039e9a0540e_bg.wasm');</script></head>
+    <script type="module">import init from '/asset/index-f1b1ef641399ddf5.js';init('/asset/index-f1b1ef641399ddf5_bg.wasm');</script></head>
     <body>
 </body></html>
