+++ backend/src/resource/sql/ddl.sql
@@ -10,13 +10,12 @@ CONSTRAINT "name" UNIQUE ("name" ASC)
 
 CREATE TABLE tag_usage (
 id INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
-blog_id INTEGER NOT NULL,
+post_id INTEGER NOT NULL,
 tag_id INTEGER NOT NULL,
-created_at INTEGER NOT NULL,
-CONSTRAINT "using_tag_UN" UNIQUE ("blog_id" ASC, "tag_id" ASC),
+CONSTRAINT "using_tag_UN" UNIQUE ("post_id" ASC, "tag_id" ASC)
 );
-CREATE INDEX blog_id_IDX ON using_tag (blog_id);
-CREATE INDEX tag_id_IDX ON using_tag (tag_id);
+CREATE INDEX post_id_IDX ON tag_usage (post_id);
+CREATE INDEX tag_id_IDX ON tag_usage (tag_id);
 -- CREATE UNIQUE INDEX blog_id_tag_id_IDX ON using_tag (blog_id,tag_id);
 
 CREATE TABLE post (
