+++ backend/src/service/asset_list.rs
@@ -1,10 +1,11 @@
 [
 ("codemirror.min.css", include_bytes!(r#"..\resource\asset/codemirror.min.css.gz"#), "text/css"),
-("index-5114c039e9a0540e.js", include_bytes!(r#"..\resource\asset/index-5114c039e9a0540e.js.gz"#), "text/javascript"),
-("index-5114c039e9a0540e_bg.wasm", include_bytes!(r#"..\resource\asset/index-5114c039e9a0540e_bg.wasm.gz"#), "application/wasm"),
+("index-f1b1ef641399ddf5.js", include_bytes!(r#"..\resource\asset/index-f1b1ef641399ddf5.js.gz"#), "text/javascript"),
+("index-f1b1ef641399ddf5_bg.wasm", include_bytes!(r#"..\resource\asset/index-f1b1ef641399ddf5_bg.wasm.gz"#), "application/wasm"),
 ("index.html", include_bytes!(r#"..\resource\asset/index.html.gz"#), "text/html; charset=utf-8"),
 ("purecss.min-b565cdb90b7aac9f.css", include_bytes!(r#"..\resource\asset/purecss.min-b565cdb90b7aac9f.css.gz"#), "text/css"),
 ("snippets/blog-frontend-3607fd663da6040e/asset/editor.js", include_bytes!(r#"..\resource\asset/snippets\blog-frontend-3607fd663da6040e\asset\editor.js.gz"#), "text/javascript"),
+("tag-4d2c4d8d06ebde6.css", include_bytes!(r#"..\resource\asset/tag-4d2c4d8d06ebde6.css.gz"#), "text/css"),
 ("toastui-editor-all.min.js", include_bytes!(r#"..\resource\asset/toastui-editor-all.min.js.gz"#), "text/javascript"),
 ("toastui-editor.min.css", include_bytes!(r#"..\resource\asset/toastui-editor.min.css.gz"#), "text/css"),
 ]
