+++ backend/src/service/server.rs
@@ -132,7 +132,7 @@ pub async fn create_warp_server(address: &str, receiver: Receiver<()>) -> Result
         .and(warp::path::end())
         .and_then(post::list);
     let tag_list = warp::get()
-        .and(warp::path("blog"))
+        .and(warp::path("post"))
         .and(warp::path("tags"))
         .and(warp::path::end())
         .and_then(tag::list);
