+++ backend/src/util/common.rs
@@ -1,7 +1,11 @@
+use std::time::{SystemTime, UNIX_EPOCH};
+
 use lazy_static::lazy_static;
 use rand::{rngs::OsRng, RngCore};
 use regex::Regex;
 use uuid::Uuid;
+
+use crate::util::result::Result;
 
 pub fn simple_uuid_with_name(name: &[u8]) -> String {
     let uuid = Uuid::new_v5(&Uuid::NAMESPACE_URL, name);
@@ -17,4 +21,9 @@ pub fn simple_uuid() -> String {
 
 lazy_static! {
     pub static ref EMAIL_REGEX: Regex = Regex::new(r"[^@ \t\r\n]+@[^@ \t\r\n]+\.[^@ \t\r\n]+").unwrap();
+}
+
+pub fn get_current_sec() -> Result<u64> {
+    let d = SystemTime::now().duration_since(UNIX_EPOCH)?;
+    Ok(d.as_secs())
 }
