+++ backend/src/util/result.rs
@@ -40,7 +40,14 @@ impl warp::reject::Reject for ErrorWrapper {}
 
 impl From<std::io::Error> for ErrorWrapper {
     fn from(e: std::io::Error) -> Self {
-        eprintln!("here 2 {}", e);
+        eprintln!("{}", e);
+        Error::ReadBlogIdDataByTagFailed.into()
+    }
+}
+
+impl From<std::time::SystemTimeError> for ErrorWrapper {
+    fn from(e: std::time::SystemTimeError) -> Self {
+        eprintln!("{}", e);
         Error::ReadBlogIdDataByTagFailed.into()
     }
 }
