+++ frontend/index.html
@@ -4,6 +4,7 @@
         <meta charset="utf-8">
         <title>70年代、80年代、90年代，怀旧、复古、回忆</title>
         <link data-trunk rel="css" href="asset/purecss.min.css">
+        <link data-trunk rel="css" href="asset/tag.css">
         <link data-trunk rel="copy-file" href="asset/codemirror.min.css">
         <link data-trunk rel="copy-file" href="asset/toastui-editor.min.css">
         <link data-trunk rel="copy-file" href="asset/toastui-editor-all.min.js">
