+++ frontend/src/val.rs
@@ -3,9 +3,9 @@ pub(crate) const USER_LOGIN_URL: &str = "http://127.0.0.1:9270/user/login";
 pub(crate) const USER_REGISTER_URL: &str = "http://127.0.0.1:9270/user/register";
 pub(crate) const USER_LOGOUT_URL: &str = "http://127.0.0.1:9270/user/logout";
 pub(crate) const USER_INFO_URL: &str = "http://127.0.0.1:9270/user/info";
-pub(crate) const BLOG_LIST_URL: &str = "http://127.0.0.1:9270/blog/list/";
-pub(crate) const BLOG_TAGS_URL: &str = "http://127.0.0.1:9270/blog/tags";
-pub(crate) const BLOG_TAG_LIST_URL: &str = "http://127.0.0.1:9270/blog/tag/";
-pub(crate) const BLOG_SAVE_URL: &str = "http://127.0.0.1:9270/blog/save";
-pub(crate) const BLOG_SHOW_URL: &str = "http://127.0.0.1:9270/blog/show/";
-pub(crate) const BLOG_IMAGE_SAVE_URL: &str = "http://127.0.0.1:9270/blog/image/save/";
+pub(crate) const BLOG_LIST_URL: &str = "http://127.0.0.1:9270/post/list/";
+pub(crate) const BLOG_TAGS_URL: &str = "http://127.0.0.1:9270/post/tags";
+pub(crate) const BLOG_TAG_LIST_URL: &str = "http://127.0.0.1:9270/post/tag/";
+pub(crate) const BLOG_SAVE_URL: &str = "http://127.0.0.1:9270/post/save";
+pub(crate) const BLOG_SHOW_URL: &str = "http://127.0.0.1:9270/post/show/";
+pub(crate) const BLOG_IMAGE_SAVE_URL: &str = "http://127.0.0.1:9270/post/image/save/";
